+++ lib/ultra_settings.rb
@@ -260,6 +260,19 @@ module UltraSettings
       @configurations.keys
     end
 
+    # Get an array of all of the configuration instances that have been loaded into memory.
+    #
+    # @return [Array<UltraSettings::Configuration>] The configuration instances.
+    # @api private
+    def __configurations__
+      @configurations.each do |name, class_name|
+        __load_config__(name, class_name)
+      end
+
+      config_classes = ObjectSpace.each_object(Class).select { |klass| klass < Configuration }
+      config_classes.collect(&:instance)
+    end
+
     private
 
     # Load a configuration class.